move time series graph into separate module
